Fail if env var is not set
